Move resource management in a separate trait
